CREATE : products 작성자 접근 권한 수정중

+++ products/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
 from .models import Product
 from .serializers import ProductSerializer
 
@@ -11,6 +12,7 @@ class ProductListAPIView(APIView) :
         serializers = ProductSerializer(products,many=True)
         json_response = serializers.data
         return Response(json_response)
+    permission_classes = [IsAuthenticated]
     def post(self,request) :
         serializer = ProductSerializer(data=request.data)
         if serializer.is_valid(raise_exception=True) :
@@ -18,13 +20,18 @@ class ProductListAPIView(APIView) :
             return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 class ProductDetailListAPIView(APIView) :
+    permission_classes = [IsAuthenticated]
     def delete(self,request,productId) :
         product = get_object_or_404(Product, pk= productId)
-        product.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if product.user == request.user :
+            product.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
     def put(self,request,productId) :
         product = get_object_or_404(Product, pk = productId)
-        serializer = ProductSerializer(product, data=request.data, partial = True)
-        if serializer.is_valid(raise_exception=True) :
-            serializer.save()
-            return Response(serializer.data)
+        if product.user == request.user :
+            serializer = ProductSerializer(product, data=request.data, partial = True)
+            if serializer.is_valid(raise_exception=True) :
+                serializer.save()
+                return Response(serializer.data)
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
